backend/pyproject.toml
[project]
name = "kuro-viewer-backend"
version = "0.1.0"
description = "FastAPI backend for Kuro Viewer"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.128.0",
    "joblib>=1.5.3",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# =============================================================================
# Ruff - Linting and Formatting
# =============================================================================
[tool.ruff]
target-version = "py313"
line-length = 140

exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
]

ignore = [
    "E501",   # line too long (handled by formatter)
    "E722",   # allow bare except
    "B008",   # allow function call in default argument (FastAPI File/Depends)
]

fixable = ["ALL"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# =============================================================================
# ty - Type Checking
# =============================================================================
[tool.ty.environment]
python-version = "3.13"
python = "./.venv"

[tool.ty.rules]
possibly-unresolved-reference = "warn"
invalid-argument-type = "warn"
possibly-missing-attribute = "warn"
unresolved-attribute = "warn"
